+++ src/Validation/SluggerFilter.php
@@ -4,19 +4,19 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Common\Validation;
 
-use Cocur\Slugify;
 use Laminas\Filter\Exception;
 use Laminas\Filter\FilterInterface;
+use Symfony\Component\String\Slugger;
 
 use function is_string;
 
 class SluggerFilter implements FilterInterface
 {
-    private Slugify\SlugifyInterface $slugger;
+    private Slugger\SluggerInterface $slugger;
 
-    public function __construct(?Slugify\SlugifyInterface $slugger = null)
+    public function __construct(?Slugger\SluggerInterface $slugger = null)
     {
-        $this->slugger = $slugger ?: new Slugify\Slugify(['lowercase' => false]);
+        $this->slugger = $slugger ?: new Slugger\AsciiSlugger();
     }
 
     /**
@@ -28,6 +28,6 @@ class SluggerFilter implements FilterInterface
      */
     public function filter($value)
     {
-        return is_string($value) ? $this->slugger->slugify($value) : $value;
+        return is_string($value) ? (string) $this->slugger->slug($value) : $value;
     }
 }